Use suite test for kubebuilder setup/teardown
