+++ .github/workflows/v2.yml
@@ -19,7 +19,7 @@ jobs:
         go-version: [1.17.5]
     steps:
       - uses: actions/checkout@v2
-      - name: Restore Go cache
+      - name: Go modules cache
         uses: actions/cache@v1
         with:
           path: ~/go/pkg/mod
@@ -27,7 +27,7 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-go-
 
-      - name: Restore node_modules cache
+      - name: Node modules cache
         uses: actions/cache@v2
         env:
           cache-name: cache-node-modules
